requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.5
django-allauth==65.3.1
django-autoslug==1.9.9
django-cities-light==3.10.1
django-htmx==1.22.0
idna==3.10
oauthlib==3.2.2
pillow==11.1.0
progressbar2==4.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python-utils==3.9.1
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.3.0
